fix composeVolumeName: remove duplicate composeVolumeName from orchestrators; add tests in cmd/upgrade
