+++ internal/orchestrators/backup_test.go
@@ -24,26 +24,6 @@ func TestBackupVolumeName(t *testing.T) {
 	}
 }
 
-func TestComposeVolumeName(t *testing.T) {
-	tests := []struct {
-		installPath string
-		suffix      string
-		want        string
-	}{
-		{"/opt/canasta/my-wiki", "db", "my-wiki_db"},
-		{"/opt/canasta/my-wiki", "mediawiki-uploads", "my-wiki_mediawiki-uploads"},
-		{"/single", "db", "single_db"},
-		{"", "db", "._db"},
-		{"/path/with spaces/wiki", "db", "wiki_db"},
-	}
-	for _, tt := range tests {
-		got := composeVolumeName(tt.installPath, tt.suffix)
-		if got != tt.want {
-			t.Errorf("composeVolumeName(%q, %q) = %q, want %q", tt.installPath, tt.suffix, got, tt.want)
-		}
-	}
-}
-
 // captureShellArg installs a mock execute.Run that records the argument
 // passed to "sh -c" and returns it. The mock succeeds without error.
 func captureShellArg(t *testing.T) *string {
